Separadas as classes de negócio de curso de extensão e as classes de negócio base do projeto. Adicionado campo de Plano de Trabalho.
