--- Code/fly/curso_extensao/templates/debug.html
@@ -1,1 +1,0 @@
-{{ value }}
